+++ apps/tx-processor/src/etransfer/index.test.ts
@@ -1,6 +1,6 @@
 // The eTransfer needs to find a file to read for it's decryption key.
 // It's not actually used, so any file will do (ie - this one...)
-process.env.DB_ACTION_PK_PATH = __filename;
+process.env.USERDATA_INSTRUCTION_PK = __filename;
 import { processUnsettledETransfers } from '.'
 import { init } from '@thecointech/firestore';
 import { GetContract } from '@thecointech/contract';