+++ libs/nft-contract/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.19](https://github.com/thecointech/thecoin/compare/v0.2.18...v0.2.19) (2021-08-21)
+
+**Note:** Version bump only for package @thecointech/nft-contract
+
+
+
+
+
+## [0.2.18](https://github.com/thecointech/thecoin/compare/v0.2.17...v0.2.18) (2021-08-21)
+
+**Note:** Version bump only for package @thecointech/nft-contract
+
+
+
+
+
 ## [0.2.17](https://github.com/thecointech/thecoin/compare/v0.2.16...v0.2.17) (2021-07-17)
 
 **Note:** Version bump only for package @thecointech/nft-contract